resolve tag/setTag conflicting method names
git-svn-id: svn://orca.physics.unc.edu/Orca64@11883 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryoView.h
@@ -20,7 +20,7 @@
 @interface ORMJDTestCryoView : ORMJDPumpCartView {
 	int tag;
 }
-- (int) tag;
+- (NSUInteger) tag;
 - (void) setTag:(int)aValue;
 - (void) drawString:(NSString*)aString atPoint:(NSPoint)aPoint;
 @end
