+++ Source/Objects/Hardware/CompactPCI/cPCI Base Classes/ORcPCICard.h
@@ -32,7 +32,7 @@
 
 #pragma mark ¥¥¥accessors
 - (Class) guardianClass;
-- (int) tagBase;
+- (NSUInteger) tagBase;
 - (NSString*) cardSlotChangedNotification;
 - (NSString*) identifier;
 - (int)  stationNumber;
