+++ Source/Widgets/Bar Graph/ORValueBar.m
@@ -122,12 +122,12 @@
 	return barColor;
 }
 
-- (int) tag
+- (NSUInteger) tag
 {
 	return tag;
 }
 
-- (void) setTag:(int)newTag
+- (void) setTag:(NSUInteger)newTag
 {
 	tag=newTag;
 }
